==> Stage 1 - Cleaning Swiss Peaks ITRA data.ipynb <==
import requests
import json
import html
import pandas as pd
import time
import numpy as np
import matplotlib.pyplot as plt
import unicodedata
import re

# Function to remove special characters
def normalize_name(name):
    # Normalize to remove accents and special characters
    name = unicodedata.normalize('NFD', name)
    # Encode to ASCII and ignore errors, then decode back to string
    name = name.encode('ascii', 'ignore').decode('utf-8')
    return name


# # URL of the JSON file
# url = "https://itra.run/Races/RaceResults/TOR330.Tor.des.G%C3%A9ants%C2%AE/2023/80226"

# # Send a GET request to fetch the JSON data
# response = requests.get(url)
# response

years = [ 
        '2024', # NO DNFs! AAAAAAAAAAAAAAAAAAh!
        '2023', # NO DNFs! AAAAAAAAAAAAAAAAAAh!
        '2022', 
        '2021',  # NO DNFs! AAAAAAAAAAAAAAAAAAh!
        '2020',
        '2019',
        '2018', 
#          '2017'
]

races = ['SwissPeaks360', 'SwissPeaks360_DUO']

#### Loading ITRA 


SP_itra_df = {}
# races = ['TOR450']

for race in races:
    for year in years:
        try:

            file = f'Data/ITRA/TXT/{race}_{year}.txt'
            df = pd.read_csv(file, delimiter='\t', 
                         dtype={'Time': 'string'})    

            # renaming columns so it is easier later to compare
            df = df.rename(columns={"Runner": "Name",
                                    "Gender": "Gender",
                                    "Nationality": "ITRA_Nationality",
                                   })

            # change the name to a title case
            df['Name'] = df['Name'].str.title()
            df['Name'] = df['Name'].apply(normalize_name)
            

            # Convert 'Time' to timedelta format
            df['Duration'] = pd.to_timedelta(df['Time'], errors='coerce')
            # Calculate total seconds and create 'Duration_Seconds' column
            df['Duration_Seconds'] = df['Duration'].dt.total_seconds()


            df['Status'] = np.where(df['Duration'].isna(), 'DNF', 'Finished')
            df['Year'] = f'{year}'
            if race == 'SwissPeaks360':
                df['Race'] = 'SP360'
            else:
                df['Race'] = 'SP360D'

    #             # Drop the intermediate columns if they're no longer needed
            df = df.drop(columns=['Time', '#'])

            df['Name'] = df['Name'].str.strip()
            df['Name'] = df['Name'].str.replace(',', '')
            df['Name'] = df['Name'].str.replace('\'', '')
            df['Name'] = df['Name'].str.replace('-', '')
            df['Name'] = df['Name'].str.replace(r"\s+", " ")

    #             print(race, year, '\n', df.head())

            # Store the DataFrame in the dictionary with a key like 'TOR330_2021'
            SP_itra_df[f'{race}_{year}'] = df
            print('* ',f'{race}_{year} {df.shape}', '\n')
        except:

            if year == '2018' :
                pass
            elif year == '2019':
                pass
            else:
                print(f'issues with {race}_{year}')
                print(f'{file}')
                print('\n')
            


for race in races:
    for year in years:
        try:
            SP_year = SP_itra_df[f'{race}_{year}']
            # Count the occurrences of each status
            status_counts = SP_year['Status'].value_counts()
            print(race, year, '\n', status_counts, '\n', '*'*40)
        except:
            pass

# Append using pd.concat
SP_itra_df = pd.concat(SP_itra_df)
SP_itra_df = SP_itra_df.reset_index(drop = True)
SP_itra_df.isna().sum()

SP_itra_df[SP_itra_df['Year'] == '2024']

def categorize_age(age):
    if age < 40: # under40
        return 'SEN'
    elif age < 50: #40-49
        return 'V1'
    elif age < 60: #50-59
        return 'V2'
    elif age < 70: #60-69
        return 'V3'
    elif age >= 70:
        return 'V4' # Over 70
#     else:
#         return str(age)

SP_itra_df['Age'] = SP_itra_df['Age'].astype('str')
SP_itra_df['Age1'] = SP_itra_df['Age'].str.replace('-', '0')
SP_itra_df['Age1'] = SP_itra_df['Age1'].astype('int')
SP_itra_df['Age Category'] = SP_itra_df['Age1'].apply(categorize_age)


## EDA of ITRA 

SP_itra_df[SP_itra_df['Gender'].isna()]

SP_itra_df['Gender'] = SP_itra_df['Gender'].fillna('M')


### Extracting Finish Category

# Convert to timedelta and get total hours (handling NaT)
SP_itra_df['Duration_Hours'] = pd.to_timedelta(
    SP_itra_df['Duration'], errors='coerce').dt.total_seconds() / 3600  # Convert seconds to hours

# Create a new column 'Finish Category'
def categorize_duration(hours):
    if pd.isna(hours):
        return 'DNF'  # Handle NaT values
    elif hours < 50:
        return '40-49 hours'
    elif hours < 150:
        num = int(hours // 10) * 10
        num1 = num +9
        return f'{num}-{num1} hours'  # Round to nearest 10 up to 150
    elif hours <= 156:
        num = int(hours // 10) * 10
        num1 = num +6
        return f'{num}-{num1} hours'  # Round to nearest 10 up to 150
    else:
        return 'Over-156 hours'
    
# Apply categorization with proper indexing
SP_itra_df['Finish Category'] = SP_itra_df['Duration_Hours'].apply(categorize_duration)
print(SP_itra_df['Finish Category'].unique())

# Define the desired order of categories
finish_category_order = ['60-69 hours', '70-79 hours',
                         '80-89 hours', '90-99 hours' ,
                         '100-109 hours', '110-119 hours',
 '120-129 hours', '130-139 hours',
                         '140-149 hours', '150-156 hours',
 'Over-156 hours', 'DNF' ]

# Set 'Finish Category' as a categorical column with the defined order
SP_itra_df['Finish Category'] = pd.Categorical(
    SP_itra_df['Finish Category'],
    categories = finish_category_order,
    ordered = True
)


# Display the DataFrame
SP_itra_df[['Year','Race', 'Duration_Hours', 'Finish Category']]

    

# grabbing all finishers 
SP_itra_df['Duration'] = SP_itra_df['Duration'].astype('str')

SP_itra_df = SP_itra_df[[
    'Name', 'ITRA_Nationality', 'Gender',  'Age', 'Age Category',
    'Duration','Duration_Seconds', 'Duration_Hours',
    'Status', 'Finish Category',
    'Race', 'Year',
]]
SP_itra_df
# SP_itra_df.to_excel(f'{race} Data/5. Clean Data for Data Visualisation/{race}_itra_including_DNF.xlsx' , index = False)



SP_itra_df.groupby(['Status', 'Year'])['Status' ].count().to_frame('count').reset_index()

### Extracting only Finishers

# SP_itra_df
SP_itra_no_DNF = SP_itra_df[(SP_itra_df['Status'] != 'DNF') ].reset_index(drop = True)
SP_itra_no_DNF = SP_itra_no_DNF[[
    'Name', 'ITRA_Nationality', 'Gender',  'Age', 'Age Category',
    'Duration','Duration_Seconds','Status',
    'Race', 'Year'
]]

SP_itra_no_DNF['Status'].unique()

SP_itra_no_DNF.isna().sum()

SP_itra_no_DNF.head()

SP_itra_no_DNF.groupby(['Status', 'Year'])['Status', 'Year'].count()

### This data might be used later data Visualisations

# grabbing all finishers 
SP_itra_no_DNF['Duration'] = SP_itra_no_DNF['Duration'].astype('str')

SP_itra_no_DNF.to_excel(f'Data/SwissPeaks360_itra_no_DNF.xlsx' , index = False)

### Extracting only 2018 and 2019 for the merge of all DNFs with scraped data

SP_itra_including_DNF_df = SP_itra_df[(SP_itra_df['Year'] == '2018') |
              (SP_itra_df['Year'] == '2019')]

SP_itra_including_DNF_df.head()

# grabbing all finishers 
SP_itra_including_DNF_df['Duration'] = SP_itra_including_DNF_df['Duration'].astype('str')

SP_itra_including_DNF_df.to_excel(f'Data/SwissPeaks360_itra_including_DNF.xlsx' , index = False)

SP_itra_df[SP_itra_df['Year'] == '2019']

for i in SP_itra_df['Finish Category'].unique():
    print(i)

SP_itra_df[pd.isna(SP_itra_df['Finish Category'])]
