==> Stage 2.2 - Scraping SwissPeaks 2023 Data.ipynb <==
# SwissPeaks Data

In this code, Swiss Peaks website is scraped. The raw data is then saved to a folder with race and year. No cleaning is done.


import requests
import json
import html
import pandas as pd
import time
import re
import unicodedata
import datetime
import random
import numpy as np
year = '2023'

# Start the timer
start_time = time.time()

## Scraping Swiss Peaks 2023 data from the website

# URL of the JSON file
url = f'https://livetrack.me/api/visitor/event_sessions/swip-{year}/races/swip-{year}-sp23-360/ranking'
print(url)

# # Send a GET request to fetch the JSON data
# response = requests.get(url)

# # Check if the request was successful (status code 200)
# if response.status_code == 200:
#     # Parse the JSON data
#     data = response.json()

#     # Optionally, save the data to a JSON file
#     with open(f'Data/JSON/SwissPeaks360_{year}_DEM_URL.json', 'w') as f:
#         json.dump(data, f, indent=4)

#     print(f"Data saved to 'SwissPeaks360_{year}_DEM_URL.json'")
# else:
#     print(f"Failed to retrieve data. Status code: {response.status_code}")

# time.sleep(delay_seconds)

# Prepare a list to hold the rows for the DataFrame
runner_info_list = []

year = '2023'
# Function to parse each entry
        # Load the JSON data from a file
with open(f'Data/JSON/SwissPeaks360_{year}_DEM_URL.json', 'r') as file:
    json_data = json.load(file)
    json_data_DATA = json_data['data']
    
    
    for data in json_data['data']:
#         print(data['enroll_participant_id'])
        
        unique_urls_for_each_runner = data["enroll_participant_id"]
        gender_category = data["enroll_participant_category"]
        gender = data["enroll_participant_gender"]
        name = unique_urls_for_each_runner.replace('swip-2023-sp23-360-', '' )
        name = name.replace('swip-2023-sp23-360d-', '' )
        name = name.replace('-', ' ' )
        name = name.title()
        print(name)
         
        all_info = [name, gender_category, gender, unique_urls_for_each_runner]         
#         # Append the row to the list
        runner_info_list.append(all_info)

initial_runner_info_df =  pd.DataFrame(runner_info, columns=['Name','Age Category', 'Gender', 
                                                'URL'])   
initial_runner_info_df

initial_runner_info_df.to_excel('Data/SwissPeaks360_2023_DEM_URL.xlsx', index = False)

## Extracting Data from JSON data

# initial_runner_info_df =  pd.DataFrame(runner_info, columns=["Name", "Age", 'Sex', 'URL'])   
# initial_runner_info_df.to_excel('Data/SwissPeaks_URL_2023.xlsx', index=False)
# initial_runner_info_df.head()

initial_runner_info_df = pd.read_excel('Data/SwissPeaks360_2023_DEM_URL.xlsx')

### Sample: Liam Vines

vines_initial_df = initial_runner_info_df[initial_runner_info_df['Name'] == 'Liam Vines']
vines_initial_df

#Where to find Vines Nationality?

year = '2023'
vines_url = 'swip-2023-sp23-360-liam-vines'
vines_stations_list = []

# URL of the JSON file
url = f'https://livetrack.me/api/visitor/event_sessions/swip-2023/enroll_participants/{vines_url}/states'
print(url)
# Send a GET request to fetch the JSON data
response = requests.get(url)

# # Check if the request was successful (status code 200)
# if response.status_code == 200:
#     # Parse the JSON data
#     data = response.json()

#     # Optionally, save the data to a JSON file
#     with open(f'Data/JSON/SwissPeaks_vines_{year}.json', 'w') as f:
#         json.dump(data, f, indent=4)

    
#     # Send a GET request to fetch the JSON data
#     response = requests.get(url)
    
#     if response.status_code == 200:
#         # Parse the JSON data
#         data = response.json()
        
#         # Extract status if available
#         status = data.get('message', 'No status message found')
        
#         # Clean and format name
#         name = vines_url.replace('swip-2023-sp23-360-', '')
#         name = name.replace('swip-2023-sp23-360d-', '')
#         name = name.replace('-', ' ').title()
        
#         print("Name:", name)
#         print("Status:", status)

#         for data_from_scrape in  data['data']['states']:            
#                 bib = data_from_scrape['bib_number']
#                 station_id = data_from_scrape['station_id']
#                 timestamp = data_from_scrape['timestamp']

#                 if timestamp is not None:
#                     timestamp = datetime.datetime.utcfromtimestamp(timestamp)
#                 else:
#                     timestamp = None  # or set a default datetime, e.g. datetime.datetime.min

#                 stay_on_station = data_from_scrape['stay_on_station']
#                 penalty_in_seconds = data_from_scrape['penalty_in_seconds']        
#                 timestamp_from_start = data_from_scrape['timestamp_from_start']
#                 timestamp_from_previous = data_from_scrape['timestamp_from_previous']
#                 time_to_first = data_from_scrape['time_to_first']


#                 all_info = [bib, vines_url, status, 
#                             station_id, timestamp, 
#                             stay_on_station, penalty_in_seconds, 
#                             timestamp_from_start,timestamp_from_previous, time_to_first]

        
#                 vines_stations_list.append(all_info)

#     else:
#         print(f"Failed to retrieve data. Status code: {response.status_code}")


            

# vines_station_df =  pd.DataFrame(vines_stations_list, columns=['Bib','URL', 'Status', 
#                                                 'Station', 'Timestamp',
#                                                 'stay_on_station', 'penalty_in_seconds', 
#                                                 'timestamp_from_start','timestamp_from_previous',
#                                                 'time_to_first'])   
# vines_station_df
        

# Merge the datsets back together
vines_df = vines_station_df.merge(
     vines_initial_df, 
    on = 'URL',
    how='left'
)

vines_df['Race'] = np.where(
    vines_df['URL'].str.contains('swip-2023-sp23-360d-', na=False), 
    'SP360D',  'SP360')

vines_df['Year'] = '2023'

vines_df = vines_df[[ "Name", "Age", 'Sex', 'Status', 
                'Station','Timestamp',
                  'Race', 'Year', 'Bib',   'URL', 
                'stay_on_station',
                'penalty_in_seconds', 'timestamp_from_start',
                'timestamp_from_previous', 'time_to_first']]

vines_df['Station'] = vines_df['Station'].str.replace('swip-2023-', '')
vines_df['Station'] = vines_df['Station'].str.replace('-', ' ')
vines_df['Station'] = vines_df['Station'].str.title()
vines_df['Station'] = vines_df['Station'].str.replace('Depart Ulrichen 1', 'Oberwald')
vines_df['Station'] = vines_df['Station'].str.replace('Depart Grande Dixence', 'Grande Dixence Bv OUT')
vines_df['Station'] = vines_df['Station'].str.replace('Arrivee Bouveret', 'Bouveret')


vines_df['Station'] = vines_df['Station'].str.replace('In', 'IN')
vines_df['Station'] = vines_df['Station'].str.replace('Out', 'OUT')

vines_df['Station'] = vines_df['Station'].str.replace('Bv', 'BV')


vines_df.head()


vines_df = vines_df.sort_values(by = 'Timestamp',  ascending = True)
# vines_df['Station'].unique()
vines_df.tail(20)

# station_order = ['Oberwald', 'Hiischi','Mosbode', 'Gere', 'Geisshitte', 'Gonerlistafel',
#        'Ueerlicherblase', 'Ulrichen R', 'Schossmatte', 'Reckingen R','Chummefurgge', 
#        'Chaeserstatt R', 'Fiesch Bv IN', 'Fiesh Bv OUT',
#        'Furggerchaeller', 'Fleschbode R', 'Lengritz R', 'Giw R',
#        'Eisten Bv IN', 'Eisten Bv OUT', 'Graechen R', 'Jungu R',
#        'Augstbordpass', 'Bluoematt R', 'Col De La Forcletta',
#        'Tsahelett R', 'Grimentz Bv IN', 'Grimentz Bv OUT',
#        'Cb Des Becs De Bosson R', 'Evolene R', 'Chemeuille R',
#        'Grande Dixence Bv IN', 'Grande Dixence Bv OUT',
#        'Col De Prafleuri', 'Grand Desert', 'Louvie', 'Plampro R',
#        'Lourtier OUT', 'Cabane Brunet', 'Cb Du Col De Mille R',
#        'Prassurny R', 'Fenetre D Arpette', 'Trient R', 'Finhaut Bv IN', 'Finhaut Bv OUT',
#                   'Col De Fenestral',   'Col D Emaney',
#        'Lac De Salanfe R',   'Col De Susanfe', 'Barme R', 'Chaux Palin R', 'Morgins Bv IN',
#        'Morgins Bv OUT', 'Portes De Culet', 'Conche', 'Tour De Don', 'Blanc Se',
#        'Taney', 'Lovenex', 'Freney', 'Debarcadere', 'Bouveret', 
                 
#        'Refuge De Chesery' # where isthis?
#                 ]

# # Set 'Station' as a categorical column with the defined order
# vines_df['Station'] = pd.Categorical(
#         vines_df['Station'],
#         categories = station_order,
#         ordered=True
#     )
    
# vines_df = vines_df.sort_values(by = 'Station',  ascending = True)
# vines_df.tail(20)

# Filter rows where 'Station' ends with ' R'
vines_df['Station'][vines_df['Station'].str.endswith(' R')].to_list()


# station_list = [['Oberwald',
#   'Ulrichen R',
#  'Reckingen R',
#  'Chaeserstatt R',
#  'Fleschbode R',
#  'Lengritz R',
#  'Giw R',
#  'Graechen R',
#  'Jungu R',
#  'Bluoematt R',
#  'Tsahelett R',
#  'Cb Des Becs De Bosson R',
#  'Evolene R',
#  'Chemeuille R',
#  'Plampro R',
#  'Cb Du Col De Mille R',
#  'Prassurny R',
#  'Trient R',
#  'Lac De Salanfe R',
#  'Barme R',
#  'Chaux Palin R',

#     'Fiesch BV IN',
#     'Fiesh BV OUT',

#     'Eisten BV IN',
#     'Eisten BV OUT',

#     'Grimentz BV IN',
#     'Grimentz BV OUT',

#     'Grande Dixence BV IN',
#     'Grande Dixence BV OUT',

#     'Finhaut BV IN',
#     'Finhaut BV OUT',

#     'Morgins BV IN',
#     'Morgins BV OUT',
  
#     'Bouveret']]


# vines_df[vines_df['Station'].isin(station_list)]

lifebases = [
    'Oberwald',
    'Fiesch BV IN',
    'Fiesh BV OUT',

    'Eisten BV IN',
    'Eisten BV OUT',

    'Grimentz BV IN',
    'Grimentz BV OUT',

    'Grande Dixence BV IN',
    'Grande Dixence BV OUT',

    'Finhaut BV IN',
    'Finhaut BV OUT',

    'Morgins BV IN',
    'Morgins BV OUT',
    'Bouveret']

lifebases_vines = vines_df[vines_df['Station'].isin(lifebases)]
# Set 'Station' as a categorical column with the defined order
lifebases_vines['Station'] = pd.Categorical(
        lifebases_vines['Station'],
        categories = lifebases,
        ordered=True
    )
    
lifebases_vines = lifebases_vines.sort_values(by = 'Station',  ascending = True)
lifebases_vines['Duration'] = lifebases_vines['Timestamp'].diff()

lifebases_vines['Duration'] = lifebases_vines['Timestamp'][lifebases_vines['Station'] == 'Bouveret'].values[0]- lifebases_vines['Timestamp'][lifebases_vines['Station'] == 'Oberwald'].values[0]

lifebases_vines[['Bib', 'Name', 'Age', 'Sex', 'Status','Duration',  'Station', 'Timestamp', 'Race', 'Year',
       'Duration']].reset_index(drop = True)

bib_start_date = lifebases_vines.loc[lifebases_vines['Station'] == "Oberwald", 'Timestamp'].values[0]
# Calculate running total duration
lifebases_vines['Running Total Duration'] = lifebases_vines['Timestamp'] - bib_start_date


lifebases_vines[['Bib', 'Name', 'Age', 'Sex', 'Status','Duration',  'Station', 'Timestamp', 'Race', 'Year',
       'Duration', 'Running Total Duration']].reset_index(drop = True)

vines_df.to_excel('Data/SwissPeaks360_VINES_2023.xlsx', index=False)

### Now extracting more runners

runners_station_list = []
failed_to_fetch = []


headers = {
    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/115.0 Safari/537.36',
    'Accept': 'application/json',
}

for unique_runner_url in initial_runner_info_df['URL'].unique():
    print("Unique runner URL:", unique_runner_url)
#             'https://livetrack.me/api/visitor/event_sessions/swip-2023/enroll_participants/swip-2023-sp23-360-andrea-dell-oro/states
    # Construct the API URL
    url = f'https://livetrack.me/api/visitor/event_sessions/swip-2023/enroll_participants/{unique_runner_url}/states'
    print("API URL:", url)
    
    # Send a GET request to fetch the JSON data
    response = requests.get(url, headers=headers, timeout=10)
    
    if response.status_code == 200:
        # Parse the JSON data
        data = response.json()
        
        # Extract status if available
        status = data.get('message', 'No status message found')
        
        # Clean and format name
        name = unique_runner_url.replace('swip-2023-sp23-360-', '')
        name = name.replace('swip-2023-sp23-360d-', '')
        name = name.replace('-', ' ').title()
        
        print("Name:", name)
        print("Status:", status)

        for data_from_scrape in  data['data']['states']:            
                bib = data_from_scrape['bib_number']
                station_id = data_from_scrape['station_id']
                timestamp = data_from_scrape['timestamp']

                if timestamp is not None:
                    timestamp = datetime.datetime.utcfromtimestamp(timestamp)
                else:
                    timestamp = None  # or set a default datetime, e.g. datetime.datetime.min

                stay_on_station = data_from_scrape['stay_on_station']
                penalty_in_seconds = data_from_scrape['penalty_in_seconds']        
                timestamp_from_start = data_from_scrape['timestamp_from_start']
                timestamp_from_previous = data_from_scrape['timestamp_from_previous']
                time_to_first = data_from_scrape['time_to_first']


                all_info = [bib, unique_runner_url, status, 
                            station_id, timestamp, 
                            stay_on_station, penalty_in_seconds, 
                            timestamp_from_start,timestamp_from_previous, time_to_first]

                runners_station_list.append(all_info)

    else:
        print(f"Failed to fetch data for {unique_runner_url}. Status code: {response.status_code}")
        failed_to_fetch.append(unique_runner_url)
        
    
    #### So we dont get blocked!!!!
    delay_seconds = random.randint(10,60)
    print('Delay before calling the server again: ', delay_seconds)
    print('*'*20)
    time.sleep(delay_seconds)



### Rearranging and saving

runner_info_df =  pd.DataFrame(runners_station_list, columns=[ 'Bib','URL', 'Status', 
                                                'Station', 'Timestamp',
                                                'stay_on_station', 'penalty_in_seconds', 
                                     # Merge the datsets back together
SP_2023  = runner_info_df.merge(
    initial_runner_info_df, 
    on = 'URL',
    how='left'
)
           'timestamp_from_start','timestamp_from_previous',
                                                'time_to_first'])   



# Merge the datsets back together
SP_2023  = runner_info_df.merge(
    initial_runner_info_df, 
    on = 'URL',
    how='left'
)


SP_2023['Race'] = np.where(
    SP_2023['URL'].str.contains('swip-2023-sp23-360d-', na=False), 
    'SP360D',  'SP360')

SP_2023['Year'] = '2023'

SP_2023 = SP_2023[["Name", "Age", 'Sex', 'Bib','URL', 'Status', 
                'Station','Timestamp',
                   'Race', 'Year',
                'stay_on_station',
                'penalty_in_seconds', 'timestamp_from_start',
                'timestamp_from_previous', 'time_to_first']]

SP_2023['Station'] = SP_2023['Station'].str.replace('swip-2023-', '')
SP_2023['Station'] = SP_2023['Station'].str.replace('-', ' ')
SP_2023['Station'] = SP_2023['Station'].str.title()
SP_2023.head()

SP_2023['Station'].unique()

SP_2023.to_excel('Data/SwissPeaks_top5_2023.xlsx', index=False)

# End the timer
end_time = time.time()

# Calculate and print execution time
execution_time = end_time - start_time
print(f"Execution time: {execution_time:.6f} seconds")
