==> Stage 2.3 - Scraping SwissPeaks 2022 Data.ipynb <==
# SwissPeaks Data

In this code, Swiss Peaks website is scraped. The raw data is then saved to a folder with race and year. No cleaning is done.


import requests
import json
import html
import pandas as pd
import time
import re
import unicodedata
import datetime
import random
import numpy as np
from datetime import timedelta

# Start the timer
start_time = time.time()

year = '2022'
url = 'https://my4.raceresult.com/184067/RRPublish/data/list?key=392b4c7157f2b40b4e5c3da0fb0c594d&listname=RESULTS%7COverall%20Results&page=results&contest=0&r=group&name=%231_Swiss%20Peaks%20360&f='


delay_seconds = 15

## Scraping Swiss Peaks 2022 Data from the website

# URL of the JSON file
print(url)

# Send a GET request to fetch the JSON data
response = requests.get(url)

# Check if the request was successful (status code 200)
if response.status_code == 200:
    # Parse the JSON data
    data = response.json()

    # Optionally, save the data to a JSON file
    with open(f'Data/JSON/SwissPeaks360_{year}_DEM.json', 'w') as f:
        json.dump(data, f, indent=4)

    print(f"Data saved to 'SwissPeaks360_{year}_DEM.json'")
else:
    print(f"Failed to retrieve data. Status code: {response.status_code}")



def convert_to_timedelta(time_str):
    """Convert time string from 'D:HH:MM:SS' to timedelta."""
    try:
        days, hours, minutes, seconds = map(int, time_str.split(':'))
        return timedelta(days=days, hours=hours, minutes=minutes, seconds=seconds)
    except ValueError:
        return None  # Handle bad format gracefully

def reformat_name(name_str):
    parts = [p.strip() for p in name_str.split(',')]
    return f"{parts[1]} {parts[0]}" if len(parts) == 2 else name_str


with open(f'Data/JSON/SwissPeaks360_{year}_DEM.json', 'r', encoding='utf-8') as file:
    json_data = json.load(file)


    # Extract the specific race data
    race_participants = json_data['data']
#     print(race_participants)
    extracted_list = []

    for participant in race_participants:
#         print(participant)
        try:
            match = re.search(r'\[img:flags/([A-Z]{2})\.gif\]', participant[4])
            country_code = match.group(1) if match else ""
            
            
                    # Convert time
            raw_time = participant[8]
            td = convert_to_timedelta(raw_time)
            formatted_time = f"{raw_time.split(':')[0]} days {':'.join(raw_time.split(':')[1:])}"

                    # Format name
            formatted_name = reformat_name(participant[5])

            extracted_entry = {
                "Bib": participant[0],
                "ID": participant[1],
                "Status": participant[2],
                "Bib_1": participant[3],
                "Nationality": country_code,
                "Name": formatted_name,
                "Gender": participant[6],
                "Birth Year": participant[7],
                "Duration": formatted_time
            }
            extracted_list.append(extracted_entry)
        except:
            print(participant)
        

# # Display extracted data
# for item in extracted_list:
#     print(item)


SP_dem_df =  pd.DataFrame(extracted_list)   
SP_dem_df['Status'] = np.where(SP_dem_df['Status'] == 'DNF', 'DNF', 'Finished')
SP_dem_df['Year'] = year
SP_dem_df['Race'] = 'SP360'
SP_dem_df['PK'] = SP_dem_df['Race'] +'_' + SP_dem_df['Year'] +'_' + SP_dem_df['ID']    

SP_dem_df = SP_dem_df[[
     'ID', 'Name','Status', 'Nationality', 'Gender', 
    'Birth Year', 'Duration', 'Year', 'Race', 'PK',
'Bib','Bib_1']]

SP_dem_df.head()

### They dont match on 3 different rows
# WHat does that mean?
# No idea
SP_dem_df[SP_dem_df['Bib'] !=SP_dem_df['Bib_1']]  

SP_dem_df.to_excel(f'Data/SwissPeaks360_{year}_DEM.xlsx', index=False)

SP_dem_df[SP_dem_df['Status'] == 'DNF']

### Using bib column to extract station 

station_list = []
failed_to_fetch= []

n = 0
# Loop through the unique bibs in the DataFrame
for unique_id in list(SP_dem_df['ID'].unique()):  # Limit to just one bib for testing

    n = n+1
    print(n, ' - ', unique_id)
    
    # URL of the JSON file
    url = f'https://my4.raceresult.com/184067/RRPublish/data/splits?key=392b4c7157f2b40b4e5c3da0fb0c594d&pid={unique_id}'
    print(url)

    # Send a GET request to fetch the JSON data
    response = requests.get(url)

    # Check if the request was successful (status code 200)
    if response.status_code == 200:
        # Parse the JSON data
        data = response.json()
        stations = data['Splits']

        # Loop through the stations
        for station in stations:
            station_name = station['Name']
            exists = station['Exists']
            
            try:
                rolling_duration = station['Gun']
                td = convert_to_timedelta(rolling_duration)
                formatted_rolling_duration = f"{rolling_duration.split(':')[0]} days {':'.join(rolling_duration.split(':')[1:])}"

            
            except KeyError:  # Use KeyError since we're checking for a missing key
                formatted_rolling_duration = np.nan  # Assign NaN if the 'Sector' key is not found

                
            try:
                aid_station_duration = station['Sector']
                
                td = convert_to_timedelta(aid_station_duration)
                formatted_aid_station_duration = f"{aid_station_duration.split(':')[0]} days {':'.join(aid_station_duration.split(':')[1:])}"

            except KeyError:  # Use KeyError since we're checking for a missing key
                formatted_aid_station_duration = np.nan  # Assign NaN if the 'Sector' key is not found


            # Store the extracted info in the station list
            extracted_info = [unique_id, station_name, formatted_rolling_duration,exists, formatted_aid_station_duration]
            station_list.append(extracted_info)
            
    else:
        print(f"Failed to fetch data for {unique_id}. Status code: {response.status_code}")
        failed_to_fetch.append(unique_id)
        
    
    #### So we dont get blocked!!!!
    delay_seconds = random.randint(5,20)
    print('Delay before calling the server again: ', delay_seconds)
    print('*'*20)
    time.sleep(delay_seconds)



# Convert the station list to a DataFrame
aid_station_df = pd.DataFrame(station_list, columns=['ID', 'Aid Station',  'Aid Station Duration', 'Exists?','Rolling Duration'])
aid_station_df['Year'] = year
aid_station_df['Race'] = 'SP360'

aid_station_df['PK'] = aid_station_df['Race'] +'_' + aid_station_df['Year'] +'_' + aid_station_df['ID']    

# Preview the DataFrame
print(aid_station_df.head())
# aid_station_df.to_excel(f'Data/SwissPeaks360_{year}_aid_stations.xlsx', index=False)

aid_station_df[aid_station_df['ID']== '1534']

aid_station_df.loc[aid_station_df['Exists?'] == False, 'Rolling Duration'] = np.nan
aid_station_df.loc[aid_station_df['Exists?'] == False, 'Aid Station Duration'] = np.nan

aid_station_df['Aid Station'].unique()

aid_station_df.to_excel(f'Data/SwissPeaks360_{year}_aid_stations.xlsx', index=False)

# End the timer
end_time = time.time()

# Calculate and print execution time
execution_time = end_time - start_time
print(f"Execution time: {execution_time:.6f} seconds")
