==> Stage 3  - SP360 ITRA 2018_2019 + SP360 DEM.ipynb <==
import pandas as pd
import numpy as np
import re

SP360_itra_including_DNF_df = pd.read_excel(f'Data/SwissPeaks360_itra_including_DNF.xlsx' )
SP360_itra_including_DNF_df  = SP360_itra_including_DNF_df.rename(columns={"ITRA_Nationality": "Nationality"}) 

SP_2020 = pd.read_excel(f'Data/SwissPeaks360_2020_DEM.xlsx')
SP_2021 = pd.read_excel(f'Data/SwissPeaks360_2021_DEM.xlsx')
SP_2022 = pd.read_excel(f'Data/SwissPeaks360_2022_DEM.xlsx')
SP_2023 = pd.read_excel(f'Data/SwissPeaks360_2023_DEM_URL.xlsx')
# SP_2024 = pd.read_excel(f'Data/SwissPeaks360_2024_DEM.xlsx')


print('columns 2020:',SP_2020.columns)
print(SP_2020.head())

print('columns 2021:',SP_2021.columns)
print(SP_2021.head())

print('columns 2021:',SP_2022.columns)
print(SP_2022.head())

print('columns 2021:',SP_2023.columns)
print(SP_2023.head())



# Create a new column 'Finish Category'
def categorize_duration(hours):
    if pd.isna(hours):
        return 'DNF'  # Handle NaT values
    elif hours < 50:
        return '40-49 hours'
    elif hours < 150:
        num = int(hours // 10) * 10
        num1 = num +9
        return f'{num}-{num1} hours'  # Round to nearest 10 up to 150
    elif hours <= 156:
        num = int(hours // 10) * 10
        num1 = num +6
        return f'{num}-{num1} hours'  # Round to nearest 10 up to 150
    else:
        return 'Over-156 hours'
# Define the desired order of categories
finish_category_order = ['60-69 hours', '70-79 hours',
                         '80-89 hours', '90-99 hours' ,
                         '100-109 hours', '110-119 hours',
 '120-129 hours', '130-139 hours',
                         '140-149 hours', '150-156 hours',
 'Over-156 hours', 'DNF' ]

for df in [SP_2020, SP_2021, SP_2022]:
    # Convert to timedelta and get total hours (handling NaT)
    df['Duration_Hours'] = pd.to_timedelta(
        df['Duration'], errors='coerce').dt.total_seconds() / 3600  # Convert seconds to hours


    # Apply categorization with proper indexing
    df['Finish Category'] = df['Duration_Hours'].apply(categorize_duration)




    # Set 'Finish Category' as a categorical column with the defined order
    df['Finish Category'] = pd.Categorical(
        df['Finish Category'],
        categories = finish_category_order,
        ordered = True
    )

    

print(SP360_itra_including_DNF_df.columns)
print('columns 2020:',SP_2020.columns)
# print(SP_2020.head())

print('columns 2021:',SP_2021.columns)
# print(SP_2021.head())

print('columns 2022:',SP_2022.columns)
# print(SP_2022.head())

# print('columns 2023:',SP_2023.columns)
# print(SP_2023.head())


SP360_itra_including_DNF_df = SP360_itra_including_DNF_df[['Name', 'Nationality', 'Gender', 
#                                                            'Age Category',
                                                           'Status', 
                                                           'Duration','Duration_Hours','Finish Category',
                                                           'Race', 'Year']]

SP_2020 = SP_2020[['Name', 'Nationality',  'Gender',
#          'Birth Year', 
                   'Status',
         'Duration', 'Duration_Hours',  'Finish Category', 'Year', 'Race']]

SP_2021 = SP_2021[[ 'Name', 'Nationality', 'Gender', 
#          'Birth Year',
                   'Status',
       'Duration', 'Duration_Hours', 'Finish Category',
         'Year', 'Race']]

SP_2022 = SP_2022[['Name',  'Nationality', 'Gender',
#                    'Birth Year',
                   'Status',
                   'Duration', 'Duration_Hours','Finish Category', 
                   'Year', 'Race']]

for df in [SP360_itra_including_DNF_df, SP_2020, SP_2021, SP_2022]:
    print(df)

SP360_dem = [SP360_itra_including_DNF_df, SP_2020, SP_2021, SP_2022]


SP360_dem_2018_2024 = pd.concat(SP360_dem )

# # Set 'Finish Category' as a categorical column with the defined order
# SP360_dem_2018_2024['Year'] = pd.Categorical(
#     SP360_dem_2018_2024['Year'],
#     categories = ['2018','2019',  '2021', '2022', '2023', '2024'],
#     ordered = True
# )


SP360_dem_2018_2024#['Year'].unique()

SP360_dem_2018_2024.groupby( ['Year', 'Status'])['Status'].count()
